+++ tools/cli-commons/src/process.ts
@@ -4,6 +4,7 @@ import buildDebug from 'debug';
 import { createInterface } from 'readline';
 
 const debug = buildDebug('verdaccio:e2e:process');
+const debugRead = buildDebug('verdaccio:e2e:line');
 
 export type ExecOutput = {
   stdout: string;
@@ -11,7 +12,7 @@ export type ExecOutput = {
 };
 
 export async function exec(options: SpawnOptions, cmd, args): Promise<ExecOutput> {
-  debug('start _exec %o %o %o', options, cmd, args);
+  debug('start _exec %o %o %o', options, cmd, args ? args.join(' ') : '');
   let stdout = '';
   let stderr;
   const env = options.env;
@@ -31,28 +32,33 @@ export async function exec(options: SpawnOptions, cmd, args): Promise<ExecOutput
   }
 
   const childProcess = spawn(cmd, args, spawnOptions);
-  // @ts-ignore
-  const rl = createInterface({ input: childProcess.stdout });
+  if (childProcess.stdout) {
+    childProcess.stdout.on('data', (data) => {
+      debugRead('data %o', data.toString());
+    });
+    const rl = createInterface({
+      input: childProcess.stdout,
+      output: process.stdout,
+      terminal: false,
+    });
 
-  rl.on('line', function (line) {
-    stdout += line;
-  });
+    rl.on('line', function (line) {
+      debugRead('line %o', line);
+      stdout += line;
+    });
+  }
 
   const err = new Error(`Running "${cmd} ${args.join(' ')}" returned error code `);
   return new Promise((resolve, reject) => {
     childProcess.on('exit', (error) => {
+      debugRead('exit %o', error);
       if (!error) {
         resolve({ stdout, stderr });
       } else {
         err.message += `${error}...\n\nSTDOUT:\n${stdout}\n\nSTDERR:\n${stderr}\n`;
-        return reject({ stdout, stderr: err });
+        const errorObj = { stdout, stderr: err };
+        return reject(errorObj);
       }
     });
   });
-}
-
-export function silentNpm(...args): Promise<ExecOutput> {
-  debug('run silent npm %o', args);
-  // @ts-ignore
-  return exec({ silent: true }, 'npm', args);
 }